Long not Int

+++ src/main/kotlin/com/classpass/moderntreasury/model/LedgerAccount.kt
@@ -29,11 +29,9 @@ data class LedgerAccount(
      */
     val ledgerId: String,
     /**
-     * an integer that is incremented every time a transaction is posted to the account. When the client is creating a
-     * ledger transaction and wants to assert that the ledger account's state hasn't changed since the last read, it can
-     * optionally pass version into the create call for each ledger entry.
+     * an integer that is incremented every time a transaction is posted to the account.
      */
-    val lockVersion: Int,
+    val lockVersion: Long,
     /**
      * Additional data represented as key-value pairs. Both the key and value must be strings. See
      * https://docs.moderntreasury.com/reference#metadata.
